backend/nlp-service/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pika==1.3.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3 